Add python script for demultiplexing
As alternative to prepare_fastq, this code uses the shell command as a
subprocess to feed into pandas dataframes for concatenation, using the
seqid as index. This would allow to run QC on the read and index pairs
separately, eg with fastp, and do the demultiplexing on the quality
reads only. We have to check

- if fastp allows filtering of complete reads based on average quality
  scores

- how we chunk the input to do in-memory processing efficiently
  (consider dask dataframes as well)