Add rule starcode for clustering barcode sequences
Purpose: We want to cluster the set of sequences that make up the
combinatorial barcodes into groups that have a small Levenshtein
distance.

Use starcode, a generic clustering tool that uses pairwise alignment
and computes the distance. It is very fast but the output file needs
to be processed properly to tag and filter the insert fastq files.